refactor: add `@vue-macros/config`
