+++ .vscode/settings.json
@@ -25,6 +25,7 @@
   "exportall.config.semis": false,
   "exportall.config.folderListener": [
     "/packages/api/src/vue",
-    "/packages/api/src/ts"
+    "/packages/api/src/ts",
+    "/packages/config/src"
   ]
 }
