+++ packages/common/src/index.ts
@@ -4,6 +4,7 @@ export * from 'ast-kit'
 export * from './ast'
 export * from './constants'
 export * from './dep'
+export * from './options'
 export * from './types'
 export * from './unplugin'
 export * from './vue'
