+++ packages/common/src/options.ts
@@ -1,0 +1,6 @@
+import type { FilterOptions } from './unplugin'
+
+export interface BaseOptions extends FilterOptions {
+  version?: number
+  isProduction?: boolean
+}
