+++ packages/config/src/config.ts
@@ -1,14 +1,14 @@
-// @ts-expect-error missing types
 import { makeSynchronized } from 'make-synchronized'
-import type { Options } from './index'
+import type { Options } from './options'
 
-export function getConfig(): Omit<Options, 'plugins'> {
+export function loadConfig(): Omit<Options, 'plugins'> {
   const url = import.meta.url
   const isDist = url.endsWith('.js')
   const workerPath = new URL(
     isDist ? './config-worker.mjs' : '../dist/config-worker.mjs',
     import.meta.url,
   )
-  const configWorker = makeSynchronized(workerPath)
-  return configWorker.getConfigAsync()
+  const { loadConfigAsync } =
+    makeSynchronized<typeof import('./config-worker')>(workerPath)
+  return loadConfigAsync()
 }
