+++ packages/config/src/define.ts
@@ -1,0 +1,7 @@
+import type { Options } from './options'
+
+export function defineConfig(
+  config: Omit<Options, 'plugins'>,
+): Omit<Options, 'plugins'> {
+  return config
+}
