+++ packages/config/tsup.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'tsup'
+import config from '../../tsup.config.js'
+
+export default defineConfig({
+  ...config,
+  shims: true,
+})
