+++ packages/macros/package.json
@@ -148,6 +148,7 @@
     "@vue-macros/boolean-prop": "workspace:*",
     "@vue-macros/chain-call": "workspace:^",
     "@vue-macros/common": "workspace:*",
+    "@vue-macros/config": "workspace:*",
     "@vue-macros/define-emit": "workspace:^",
     "@vue-macros/define-models": "workspace:*",
     "@vue-macros/define-prop": "workspace:^",
