+++ packages/macros/tsup.config.ts
@@ -1,7 +1,1 @@
-import { defineConfig } from 'tsup'
-import config from '../../tsup.config.js'
-
-export default defineConfig({
-  ...config,
-  shims: true,
-})
+export { default } from '../../tsup.config.js'
