+++ packages/shim.d.ts
@@ -5,3 +5,12 @@ declare module '*?raw' {
 interface ImportMeta {
   DEV: boolean
 }
+
+declare module 'make-synchronized' {
+  type AwaitModule<T> = {
+    [K in keyof T]: T[K] extends (...args: infer A) => infer R
+      ? (...args: A) => Awaited<R>
+      : T[K]
+  }
+  export function makeSynchronized<T>(module: URL): AwaitModule<T>
+}
